requirements.txt
pixivpy
Pillow
Flask
Flask-Cors
